Update patch set 1
Patch Set 1: Code-Review-1

Maybe you're already anticipating this, but I think there will be an issue with encrypted userdata partitions.  blkid returns empty for mine.

It seems the recovery doesn't easily support encryption anyway, but we wouldn't want to wipe their /data without warning.  Perhaps I am not understanding the context of when this is used, though.  Is it for every flash?

Patch-set: 1