Update patch set 1
Patch Set 1:

> shouldn't be required once the lzma patch is merged :)

That's assuming we're ready to make changes to other recoveries. I'd much rather have this than to be forced to change the recovery compression method for each recovery project I work on.

The change is recommended by the TWRP man himself, so there's that. It's small, and won't hurt anyone -- on the contrary, it will benefit a large chunk of CM end-users.

Patch-set: 1