Update patch set 2
Patch Set 2:

> sorry but this is not acceptable
 > reduce recovery instead of increasing RD

Not acceptable!?
There are recoveries out there that actually work, as opposed to being the software equivalent of a chihuahua with Down syndrome and no legs. As a device maintainer you should take that into account, because this is effectively forcing users to roll with Cyanogen Recovery, and Cyanogen recovery alone.
At least tell us _why_ this is not up to your expectations. What are the cons you see in this that do not meet your standards, if you will?

Or are you going to just post a -2 and let this patch die off?

Patch-set: 2