Update patch set 2
Patch Set 2:

Oh we're pasting links now? I guess answering questions isn't something the cyanogenmod gerrit is known for. While we're sharing links left and right, let me introduce you to a very good friend of mine, the z3c:
https://twrp.me/devices/sonyxperiaz3compact.html
http://forum.xda-developers.com/z3-compact/orig-development/recovery-twrp-2-8-7-0-18-8-2015-t3181623/post62528872#post62528872 Note in this thread the lack of users reporting it working, such as http://forum.xda-developers.com/showpost.php?p=62400419&postcount=8 , http://forum.xda-developers.com/showpost.php?p=62423501&postcount=9 and http://forum.xda-developers.com/showpost.php?p=62396034&postcount=4 .

There are numerous reports of the whole thing being a huge screwup on CyanogenMod and ROMs whose maintainers have not merged some crucial changes. The MultiROM dev even went as far to compress his version of the recovery (which is in essence TWRP modified to support MultiROM) with LZMA, but again people cannot use that because CM's ramdisk does not support that kind of extraction.

You hold a grudge against TWRP for a reason, that much is evident. Why you do not permit others to use it and fix the issues that have been screaming in our faces for several weeks now, or why even asking about it results in the kindergarten-level "silent treatment" is beyond me. And a bunch of other people, many of whom are _much_ more credible than I am, but are too tired of this kind of treatment to keep on pushing.

Patch-set: 2