Update patch set 2
Patch Set 2:

Brin Taylor,
i agree with giving more respect.
Yes we want to use FOTAKernel and this commit is for using it. With only 8MB it is verry hard to get a full working COTR or TWRP. How about ~10MB for buffer? Thats only 2MB more but 2MB means a lot to us. Is that a possibility?

Patch-set: 2