Update patch set 2
Patch Set 2:

I do apologize for my lack of politeness and everything I said; I agree it could have been said in a more fitting way. However, I was not the only one whose way of communication was perceived as insulting by the opposing party -- however, everything I said, though in a very unprofessional manner, still holds true. Answers and arguments are ignored completely, empty promises are made, and the lack of communication is quite shocking, to say the least.

Brin, while you report not having issues, the opposite is obvious from the few links I posted in my anger.
Saying that you do "not support the recovery" is as unfair as can be to the end user, because your decision is limiting a large number of people in their options. I was once convinced by a CM maintainer that extract_elf_ramdisk was made in hopes of giving the user a choice -- one that you seem to be very eager to take away.

I mean no disrespect when I say this, Giulio, but ignoring replies should not be an acceptable way of handling a conversation.
I appreciate the fact you at least made a promise to test (though the -1 is a tad confusing), and I can only hope you keep it in a more honorable way than a certain comrade of yours did his.

Patch-set: 2