Update patch set 2
Patch Set 2:

> Jari, most of your statements are completly wrong and i guess u
 > dont understand how complex is test around 20 devices, anyway, my
 > -1 is mainly because someone promised me to review this commit and
 > also because comment is wrong (patch increase to 20MB comment says
 > 13MB)

I would greatly appreciate a comment on what I said that was so very wrong. This whole recovery in boot thing is not very well documented, and all my knowledge on it is based off the code I understood and the testing I did on my two Sony devices (nozomi back in 2013/14 and z3c), and short conversations with TWRP and omnirom developers. I wasn't questioning your testing procedure -- on the contrary, I was thanking you for making a promise to do it.
The commit message can easily be corrected, so you could have pointed that out as something we missed and should fix.

Patch-set: 2