Update patch set 3
Patch Set 3: Code-Review-1

Let's just abandon this patch? I pushed it to CM (cm-11.0) last year, and abandoned it because Giulio said TWRP may become bigger some day. And if you ask me, that's true. This patch will work today, maybe even tomorrow, but as you see, it's not a long-term solution (we can't go on and increment the buffer on a regular basis).

I'd compress the recovery ramdisks with LZMA; I had to do that very often in the past for other legacy devices with small partitions.

Patch-set: 3
Label: Code-Review=-1