Update patch set 3
Patch Set 3:

That's my understanding as well -- the FOTAkernel partition can have a compressed recovery (though again, we'd need a way to uncompres it at boot) that's up to 16 MB big. But the buffer applies to the size of the uncompressed recovery ramdisk (which also implies that the FOTAkernel partition doesn't store the uncompressed ramdisk). If my observations are true then this patch does make sense.

Do correct me if I'm wrong, as I wrote earlier I still have a long way to go with studying the boot system.

Patch-set: 3