Update patch set 3
Patch Set 3: Code-Review-1

Path log which could be interesting.

KEYCHECK_TIMEOUT logic has to remain as is.
Where would /init.real come from ?

I like the idea of a "Power & Volume-" boot to recovery, but from the test I made, there's no special warmboot value set on such a boot, therefore no detection of this boot from what I saw so far.

As for these changes, can't see anything related to this keys boot detection.

Have a look to the FOTA update commit, was planning on doing it, it'll fix the issues on empty FOTA.

Patch-set: 3