Update patch set 5
Patch Set 5:

the two commits build correctly and disable the keycheck correctly, but the -c option is still not working, it does not load the recovery from FOTA and so continues to boot as normal.

If I add back the DEV_STOCK_CHECK part of my last commit, it will boot to recovery correctly.

I am not sure if the -c has ever worked correctly, we removed it in the init.sh at Mozilla, probably because it was broken.

Patch-set: 5