Update patch set 1
Patch Set 1:

(5 comments)

Since I started this sony_init I had as an objective to keep it as open & configurable / adaptable to each device as possible, so a feature addition (and here an access denial) would still be better to be changeable on device side.

These all are presets configs of how it's going by default, and if any device maintainer one day wants to do it, because he found a way to mount what's needed (I went into it but really didn't like the idea) and wants to use it, it remains accessible.

That's my point of view on this shared init, but might not be shared.

Patch-set: 1