Update patch set 1
Patch Set 1: Verified+1

Works fine for me, with and without root only /system/bin/toybox is returned by "which toybox".

Lightest solution to avoid per-device renaming of the toybox binary and avoid a possible binary mismatch.

Basically doesn't matter on CyanogenMod because it's the same binary sources, but on AOSP and Stock ROMs using the init, there's mixes between our toybox for the bootimage and the normal toybox sometimes that might introduce issues.

An other issue-case would be if a user flashes an older bootimage for any reason, and would therefore use an "older" toybox for all root commands.

Patch-set: 1