Update patch set 1
Patch Set 1:

Alexandar, from a quick look, I would ask you to split this into two changes to make it easier.
One for the code style changes, and then another for the issue you're trying to resolve. Makes it clearer, easier to review and faster to merge,

I'll leave the rest to Adrian, since he has been the main person behind this.

Patch-set: 1