Update patch set 1
Patch Set 1:

I started looking into it, on the principle it's a good idea, I like it.

As Chirayu mentioned, most of current active CM devices actually need it entirely, and that won't change.

I'm ok if some devices want to make it optional, but that doesn't concern the 2011-2012-2013 families.

Now on a more changes oriented aspect, I would like to point out that you should be careful when doing changes to files on the formats and styles used. The init_board_common.h file was converted to a Windows ending format (CRLF), instead of the others staying in Unix (LF), hence the whole file appeared "changed".

Also avoid tabulations, 4 spaces is the standard.
Careful with leftover white spaces too, shown in red on gerrit,
and the 80 chars limit for line breaks.

The finish_init should not be filled by default for all devices,
these functions are meant to be overridden by each device if needed,
for example finish_init is mostly used to turn off LEDs & vibrator.

Then on the custombootimg side, there's a big problem,
you're renaming the init from the recovery image to match your precise needs.

Except that's only for the inbuilt CM recovery & will break
every FOTA recovery that still has the "init" name (TWRP, old CM),
and will also break the bootability of the CM Recovery image released with ROMs.

Finally, I really dislike the logic of "excluding" things,
after doing a lot of testing & checking of the cleanups,
I just found out it wiped recursively my data & storages on my mint,
my fault since I was doing the tests but shows how dangerous it can be,
a listed cleanup of the files to remove is more reliable & understandable.

I also see this is going to break XperiaMultiROM dual ramdisk support so will check that later too and make the needed changes with Christian probably.

Patch-set: 1