Update patch set 2
Patch Set 2:

Hi Adrian,

thanks for the cleaner implementation. I had some trouble with the remote connection to my build server such that I had to fix this problem first, before I had some time to work on the "C-style" implementation today. Your version came in time such that I did not invest too much time into this.

Also, thank you for your general comments! The mess in the changes was a result from setting up a proper dev environment, but this problem is solved on my side now. Sorry for that.

Regarding the re-implementation, all the points your are mentioning do make sense in my opinion. The init renaming for the recovery is actually not needed, only for the standard android boot. If the recovery is booted, the init symlink is overridden anyways (also applies to the FOTA init).

Regarding the removal of non-needed files: The whitelisting approach also seems safer to me. There were two reasons that made me choose the blacklising approach:
1) If some additional files were added to the ramdisk (that does not belong to the needed set) they would be deleted automatically since the required adjustment in the file list can be easily forgotten and
2) the environment should be somewhat controlled as the system is in a very early stage where none of the device filesystems should be accessible.
Hence, I am really wondering how this could have wiped out your data, but if this happened, it is absolutely enough reason to drop the placklisting approach. I think, that its disadvantages are neglible over the data safety of the whitelisting approach.

Further, the required changes for the multirom portion could be also done by me (I am maintaining Mrom for the Z2). The required change would be to make Mrom recognize the renamed init binary. The "standard" layout should not be a problem since XperiaMultiRom is capable to handle the standard and the enclosed ramdisk layout (if I remember the code correctly). However, I think it would be good to include Christian in the discussion as he has a better overview than me.

Finally, thank you for your work Adrian! I like the implementation from a first quick look. I will test it now and report back with the results.

Patch-set: 2