Update patch set 2
Patch Set 2: Code-Review+1

So, I have finished testing and all is operating as desired: Booting CM, the FOTA recovery, and the enclosed recovery.

I also tested mulitrom (only booting the primary rom) and it is also working as expected. Seems to be of previous work for Sony AOSP roms that already did the init.real renaming, but this is pure guessing.

Further, I did a code review of the change and I could see nothing that would prevent merging. I think it is a good idea to introduce the wildcard file remover and using the existing dir_unlink_r method where needed.

Patch-set: 2