Update patch set 2
Patch Set 2:

I am ok with the changes, so far all tests look fine to me.

However, I need confirmation that latest SuperSU 2.76 does work fine.

Support for Sony 8x60 with ELF bootimages is not present for the moment in SuperSU,
as well as the "name" of my kernel partition : 'Kernel' (Not a joke :D).

So basically, the same kind of port that I did to MultiROM libbootimg
would need to be done to SuperSU installer to support unpacking,
editing and repacking a proper bootimage with Sony ELF format.

Therefore, SuperSU 2.52 remains the one to use on our AOSP-based ROMs,
but newer device with a regular bootimage should be working just fine now.

Patch-set: 2